T6 2.0.0

+++ lib/AFE-Data-Structure/AFE-DHT-Sensor-Structure.h
@@ -20,7 +20,7 @@ struct DH {
   uint8_t type;
   TEMPERATURE temperature;
   HUMIDITY humidity;
-  uint16_t interval;
+  uint32_t interval;
   boolean sendOnlyChanges;
   boolean publishHeatIndex;
   boolean publishDewPoint;
